+++ pb_public/flutter_bootstrap.js
@@ -38,6 +38,6 @@ _flutter.buildConfig = {"engineRevision":"13e658725ddaa270601426d1485636157e38c3
 
 _flutter.loader.load({
   serviceWorkerSettings: {
-    serviceWorkerVersion: "2880655068"
+    serviceWorkerVersion: "2163173537"
   }
 });